Make PTC peek the queue before consuming
This essentially makes queue consumption the last thing that happens,
rather than the first, so we stand a lower chance of losing data and
have a higher chance of duplicating (which is fine, WFL can handle
that downstream).

The test just uses a locally instantiated queue but it does demonstrate
that the peeking works. Notably, the existing integration test required
no changes, since the observable behavior of a normal run is basically
unchanged.