Fix our jar building
The comma in `(compile,'ptc.start)` is because the actual code that needs to be evaluated is `(compile 'ptc.start)` but the deps.edn file is parsed such that the space will split the latter half into another argument, causing an issue as then the code can't be read (you'll just get really cryptic EOF issues). Commas are universal whitespace in Clojure, which I hate, but here I suppose it is a convenient workaround for Clojure having weird parsing of its own dependency file.